requirements.txt
typer>=0.12
Pillow>=10.0
numpy>=1.26
requests>=2.32
transformers>=4.53.0
torch>=2.4.0
torchvision
Pillow
fastapi
uvicorn[standard]
python-multipart
huggingface_hub
timm
peft
unsloth
datasets
